pyproject.toml
[build-system]
requires = ["setuptools>=80.10.2", "wheel>=0.46.3"]
build-backend = "setuptools.build_meta"

[project]
name = "ragonometrics"
version = "0.1.0"
description = "Simple RAG pipeline for economics papers"
requires-python = ">=3.11"
dependencies = [
  "openai>=2.17.0",
  "anthropic>=0.39.0",
  "streamlit>=1.54.0",
  "requests>=2.32.5",
  "faiss-cpu>=1.13.2",
  "psycopg[binary]>=3.2.12",
  "psycopg_pool>=3.2.6",
  "sqlalchemy>=2.0.44",
  "alembic>=1.17.1",
  "rank_bm25>=0.2.2",
  "numpy>=2.4.2",
  "pydantic-core>=2.41.5",
  "pydantic>=2.12.5",
  "tqdm>=4.67.1",
  "fpdf2>=2.8.3",
  "pdf2image>=1.17.0",
]
  

[project.optional-dependencies]
# Developer/test dependencies
dev = [
  "pytest>=9.0.2",
]

# Optional OCR/image dependencies (install only if you need OCR fallback)
ocr = [
  "pillow>=12.1.0",
  "pytesseract>=0.3.13",
]


[project.scripts]
ragonometrics = "ragonometrics.cli.entrypoints:main"

[tool.setuptools]
packages = { find = { where = ["."], include = ["ragonometrics*"] } }

